Move helper function to telemetry pkg
